+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <link rel="stylesheet" href="styles.css" />
+    <link rel="stylesheet" href="styles/styles.css" />
 
     <!-- FONTS -->
     <link rel="preconnect" href="https://fonts.gstatic.com" />
@@ -16,6 +16,8 @@
     <title>😎️ Rock, Paper Scissors 😎️</title>
   </head>
   <body>
+  <section>
+
     <h1 class="greeting">Hi</h1>
     <h1 class="title">Let's play Rock,Paper, Scissors!</h1>
     <div class="round">
@@ -56,7 +58,7 @@
         </div>
       </div>
     </div>
-
+  </section>
     <script type="module" src="index.js"></script>
   </body>
 </html>
