+++ index.ts
@@ -1,7 +1,5 @@
-import Game from "./Game.js";
-import {btns} from "./selectors.js";
-
-// Start the Game
+import Game from "./Classes/Game.js";
+import {btns} from "./selectors/selectors.js";
 
 const userName = prompt(
     "Nice to meet you! My name is Max and I created this simple Rock, Paper, Scissor game. \n What's your name? "
